Extract the test methods
