+++ android-pdf-viewer/src/main/java/com/github/barteksc/pdfviewer/annotation/core/shapes/Rectangle.kt
@@ -17,36 +17,6 @@ data class Rectangle(
     val relations: Relations? = null,
 ) : Shape(type, points)
 
-fun getMockedRectangleCorners() = listOf(
-    PointF(30.606155F, 22.65048F),
-    PointF(60.60616F, 22.65048F),
-    PointF(60.60616F, 62.65048F),
-    PointF(30.606155F, 62.65048F)
-)
-
-fun getMockedRectangleRelations(): Relations {
-    val documentation = Documentation(16, "583")
-    val documentations = listOf(documentation)
-    return Relations(documentations)
-}
-
-fun getMockedRectangle(): List<Rectangle> {
-    val corners = getMockedRectangleCorners()
-
-    val edges = listOf(
-        Edge(corners[0], corners[1]),
-        Edge(corners[1], corners[2]),
-        Edge(corners[2], corners[3]),
-        Edge(corners[3], corners[0])
-    )
-
-    val relations = getMockedRectangleRelations()
-
-    val rectangle = Rectangle(points = corners, edges = edges, relations = relations)
-
-    return listOf(rectangle)
-}
-
 fun List<PointF>.generateRectangleEdges(): List<Edge> {
     val edgeTopHorizontal = Edge(this[0], this[1])
     val edgeRightVertical = Edge(this[1], this[2])
